whisper_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
numpy==1.24.3
soundfile==0.12.1
